Addressing Oliver's comment on commit 8b11609
    In the thelma.db.mappers.species module, the ordering of the imports
is not alphabetical. If you press Shift-Cmd-O in PyDev inside a module,
the imports are automatically arranged; we should always use this
auto-arrange mode to minimize spurious diffs;
    You forgot to update the doc strings for the
thelma.db.schema.tables.tissue and thelma.models.tissue modules. You do
not need to put an acronym there; just a brief description of what the
code in the module is doing;
    In the thelma/resources/configure.zcml, change at line 3116ff, you
are using the representation name of the cell_line attribute of the gene
resource (i.e., cell-line instead of cell_line). This should raise an
error with the latest everest and will have no effect in earlier
versions. Apart from that, you probably do not need to configure the
attribute anyways since the default is to represent member attributes as
a link, which is presumably what you want here;
    Again in thelma/resources/configure.zcml, change at line 3425ff, the
entity reference is wrong (should be thelma.models.tissue.Tissue, not
thelma.models.tissue.Tissues). Again, you are using cell-lines instead
of cell_lines in the attribute configuration section;
    In the thelma.resources.tissue module, you should use tissues, not
tissue, as the root name of the collection (should always be the plural
form);
    In the XML schema, you forgot to replace SpeciesCollectionLinkType
with TissueCollectionLinkType :-) ;